DashCustomItem "prevent reuse" rework
Instead of using unique reuse identifiers for every DashCustomItem, the custom item views will be cached and only reused for the item. they were created The container cells will be recycled but the cached view will be added to the container cell from the cache (and detached from the previous container cell).